requirements.txt
flask
line-bot-sdk==1.20.0
pandas
openpyxl
requests
python-dotenv
google-generativeai
